tested some meta info

+++ src/org/naivedb/utils/FileUtils.java
@@ -6,6 +6,7 @@ public class FileUtils {
 
     // this function is extremely dangerous, be careful
     public static void deleteAll(File file) throws IOException{
+        if (!file.exists()) return;
         if (file.isFile() || file.list().length == 0) file.delete();
         else {
             for (File f: file.listFiles()) {
